Fix updating ui resources after creating new form
Before we have tried to update resources, however
it was done too early and ui file, which
has been created, was not inserted to any project yet.
We were connected to ProjectTree::filesAdded signal
and apparently whenever it is sent the added ui file
is still not inserted to the project, thought
the signal informs about that fact.

Now we connect to every project's fileListChanged() signal instead.

Task-number: QTCREATORBUG-15560
Change-Id: I11883a0d87fdc7c46897cef68466c1038dbaefb8
Reviewed-by: Tobias Hunger <tobias.hunger@qt.io>