Doc: Move vcs-system specific sections to separate topics
This is done in preparation for moving some Qt Creator
plugins and their documentation into their own repositories.

All the changes have to do with the structure of the documentation.
There are no content changes, except removing headings or changing
them to better fit the new structure and adding some leading
sentences or links.

Change-Id: I1499187121a1a1cc3c6b635801fb540ac439fd0b
Reviewed-by: Jarek Kobus <jaroslaw.kobus@qt.io>
Reviewed-by: Orgad Shaneh <orgads@gmail.com>
Reviewed-by: Hugues Delorme <delorme.hugues@fougue.pro>