Theme: allow taking implicitly shared copies of the creator theme
We want to install the theme as singleton in various QML engines.
QML engines take ownership of their singletons and want to delete them
on exit. In order to accommodate for this we need copies of the theme.
As all the containers in ThemePrivate are implicitly shared, it
shouldn't be a lot of effort to copy it.

Change-Id: I52391add2fe8f540c19c804945ce9d9161905b90
Reviewed-by: Thomas Hartmann <thomas.hartmann@qt.io>