CMake: Always put in a CMakeLists.txt file
Even if there was none found by the project reader. This can happen
when the cmake binary can not be started at all with the server-mode
reader.

I wanted to avoid just creating this node as that does mess with the
expansion state of the tree. But considering that we know that the
tree is empty otherwise that should actually not be a problem.

Change-Id: I5ae5ed3e2fa7eebdfc1f9b7d55d70417abfcd125
Reviewed-by: Alessandro Portale <alessandro.portale@qt.io>