QmlDesigner: Do not throw in destructor of transaction
In C++ noexcept(true) is the default for destructors.
Instead of poluting the code base with noexcept(false),
we catch the exception inside the destructor of the transaction.
Typically the application code does not anything special to handle the
error, so showing the error message is good enough.

If the application code wants to handle exceptions commit()
has to be called instead of relying on RAAI.

Change-Id: Ic73cacb197737698454eca9270d07fd897c97ca8
Reviewed-by: Tim Jenssen <tim.jenssen@qt.io>