Fix off-by-one error which stops qtchooser finding qmake
At startup,  qtcreator prints an error:

% qtcreator
"The command \"/qmake\" could not be started."

This appears to be because of an off-by-one error in
BuildableHelperLibrary::qtChooserToQmakePath(const QString &path) which
parses the output of a call to "qtchooser -print-env" to extract the
value of the QTTOOLDIR variable.  The code attempts to move past the
prefix string to find the end quote, but by moving one too few
characters it finds the starting quote instead, resulting in a zero
length string result.

Change-Id: I74368f10a81eda2286ae735bdc595c0f92e4e665
Reviewed-by: Tobias Hunger <tobias.hunger@qt.io>
Reviewed-by: Eike Ziller <eike.ziller@qt.io>