Timeline: Improve display of time labels
As time labels are formatted with a precision that matches the context
we are generally interested in the right part of the string if it
doesn't fit into the cell. Therefore, elide left. Also, the label
corresponds to the line left of it, so put the margin to the right,
rather than to the left.

Change-Id: I953b1bc674e131dbfcf873774127c6360163f20f
Reviewed-by: Christian Kandeler <christian.kandeler@qt.io>