QmlDesigner: Fixing regression for id renaming
When we rename an id we actually trigger a refactoring
operation on the qml text.
Due to recent changes we delay and compress qml changes
from the text side using a 400ms timer.
This is fatal for renaming ids.
When we trigger refactoring operations on the text we have to ensure
the update is done instantly without delay.

To ensure this we use the memeber m_instantQmlTextUpdate.

Change-Id: Ie007ee99f201cd444033c38630d7e7a9f5a673b5
Reviewed-by: Tim Jenssen <tim.jenssen@qt.io>