Change "Executing in" to "Running in"
While I understand that the term "execute" is techically accurate, to
execute has many meanings, not all of them are positive. Execute can
also mean to kill someone. We should try to have more geopolitically
neutral terms in an IDE like QtCreator since it is used all over the world.

Change-Id: Icd6742865afa7a4469764e38eee961d7a96c7d5f
Reviewed-by: Friedemann Kleint <Friedemann.Kleint@qt.io>
Reviewed-by: Leena Miettinen <riitta-leena.miettinen@qt.io>