QtSupport: Use binary string from Qt's libraryLoader to refine ABI
Use binary string from Qt's library loader class to refine ABI if that
is not complete yet. This is necessary to reliably find the compiler
used in static Qt builds on windows (using MSVC compilers). Before we
were able to rely on the mkspec to pick a working compiler, that does no
longer work with Qt 5.8.

Task-number: QTCREATORBUG-17534
Change-Id: I50c428cbaa986a26ba165cff499f9e0be3937ea2
Reviewed-by: Tim Jenssen <tim.jenssen@qt.io>