fix fallbacks in external property parsing, take 2
while the fallback chain really *is* "Put" => Raw => Get => Src, this
implies that to get sane looking condensed qmake output with only the
plain and the /raw property variants, the /get and /src variants must be
lacking the sysroot, which is profoundly useless in the case of an
installed prefix build with sysroot and without extprefix.

so to get actually useful paths, we just drop Raw from the fallback
chain. qmake will be fixed later to actually conform with this much
saner interpretation.

amends 61419e7bf.

Change-Id: I76fced03c5774ad581df2bf6ccc92a908fa243ca
Reviewed-by: Christian Kandeler <christian.kandeler@qt.io>