QmlDesigner: Implementing actions for stacked container
* Adding action for adding item
* Adding action for decreasing index
* Adding action for increasing index

User visible strings for the actions are defined in:
componentcore_constants.h

The name of the index property is given by a hint in the .metainfo file.
If there is no hint we try "index" and "currentIndex".

This is tested with StackedLayout but should work with similar
containers like SwipeView.

Change-Id: Ief1b4206db1950130bcb0edcbd61885b3ec9d8b0
Reviewed-by: Leena Miettinen <riitta-leena.miettinen@qt.io>
Reviewed-by: Tim Jenssen <tim.jenssen@qt.io>