QmlDesigner: Refine painting of place holder
Generally painting the place holder for invisble items makes
sense.
The logic is different for items stacked containers,
since those are overlapping and we are only interested
in the current visible one.
Also it does not make sense to paint the placeholder if already
the parent is invisible, since this just creates noise.

The logic is now more complex, but the usabilty increases significantly.


Change-Id: Ia25f2877c79cace475f0b99fd11ebd242c96e4bb
Reviewed-by: Alessandro Portale <alessandro.portale@qt.io>