QmlDesigner: Improve performance of nodeAtTextCursorPosition
We find candidates by a linear search using just the offset.
This way we have to calculate the length (which is slow) only for
a very few selected nodes.
The number of nodes we have to calulate the length for is bounded by
the maximum nesting, which is for any sane QML file smaller then 10.
Before we had to calculate the length for every node.

Change-Id: I53a03818e9986db4a224aef876b333444aa1c15b
Reviewed-by: Tim Jenssen <tim.jenssen@qt.io>