CppEditor: Use qFuzzyCompare when comparing member variable for qproperty refactoring
For a line like

    Q_PROPERTY(qreal foo READ foo WRITE setFoo NOTIFY fooChanged)

the generated setter will now use "if(qFuzzyCompare(m_foo, foo))" instead of "if (m_foo == foo)"
for types that are supported by qFuzzyCompare (that is: qreal, double and float).

A warning stating that "Floating point comparison needs context sanity check" is as well added
to remind the user to check/fix the generated code.

Change-Id: I8d274d5c072d107f0d04b1e153b5cc183e6317fc
Reviewed-by: Nikolai Kosjar <nikolai.kosjar@qt.io>