Gerrit: Improve curl detection
Until Git for Windows 2.12.0, curl was shipped twice - a MinGW version
in mingw{32,64}/bin, and MSYS2 version in usr/bin[1].

On 2.12.0, the MSYS2 version was removed, leaving only MinGW.

The plugin only searches for curl in usr/bin, so it is clearly broken
with 2.12.

There is no reason to add mingw*/bin to GitClient::gitBinDirectory()
(and return a list), because there are no other useful tools there. The
other tools that use gitBinDirectory are patch and ssh, which are both
in usr/bin.

[1] https://github.com/git-for-windows/git/issues/1069

Change-Id: I5eb5fa727fa384835792c59fd018fdfa31594927
Reviewed-by: Tobias Hunger <tobias.hunger@qt.io>