Beautifier: Fix formatting with AStyle and make it more robust
Only version 2.04 added an extra newline. Later versions are fixed, so
the should not remove a newline.
In addition if piping is enabled we stream "\n" line endings so tell
AStyle this explicitly by passing -z2. This disables the automatic
detection which can be faulty with short code fragments.

Task-number: QTCREATORBUG-17546
Change-Id: Ib90526234bfbc4d3530fd9c3a15bdfb3af02bebd
Reviewed-by: David Schulz <david.schulz@qt.io>