QmlDebugger: Limit recursive property lookups
The QML engine will send the number of children in an object as the
"value" property together with the ref. We can use that to determine
if an object has children without actually fetching it.

Also, always save the type of a ref, even if we don't know the value
yet. This avoids redundant lookups.

Task-number: QTBUG-42435
Change-Id: Ia423b661187dc054fb5782349d1256711ee11265
Reviewed-by: hjk <hjk@qt.io>