QmlDesigner: Fix nodeAtTextCursorPosition()
The old implementation assumes that the order of nodes
returned by directSubModelNodes() is the order in the QML\
text document. This is not always true if we mix different
node properties.
If we cannot rely on the sorting the algorithm becomes slow.

This fixes selection of delegate items in styles.
e.g.
Button {
   text: Text {
   }
   some children
}

The selection of the Text element did not work in all cases before.

Change-Id: Icf89bdcab4cb9d2a9fe395d1fbf4e961c09e10e1
Reviewed-by: Tim Jenssen <tim.jenssen@qt.io>