QmlDesigner: Fix performance regression when selecting
The method nodeContainsCursor() is rather slow for huge documents
and we do not really have to call it. Checking if the selected node
does contain the cursor is enough and can be done in constant time.
Since we do this whenever the selection changes this fixes notable slowdown
in huge documents.

Change-Id: Ie525f0fbcacd6abb72f843f59a793c4f82190a0c
Reviewed-by: Tim Jenssen <tim.jenssen@qt.io>