RemoteLinux: Always kill the whole process group
This way we can get rid of processes spawned from an initial shell, e.g.
in case of the perf profiler. Also, if your target application spawns
additional processes, we probably want to kill those. The kill -pgid
only works if the target application is actually the process group
leader. So, in case it isn't, also kill it individually by pid.

Change-Id: Ifcaf3764d21281c3cac107419f84f6bc527854da
Task-number: QCE-91
Reviewed-by: Tobias Hunger <tobias.hunger@qt.io>