ScxmlEditor: Retain explicitly set "initial" attributes
If the user sets a state to be initial and then moves it around without
reparenting it, the initial attribute has to be retained. This is
achieved by saving a "removedInitial" attribute in the editor info when
the state is released from its parent and restoring the attribute when
it is reattached. This even works if the state is temporarily reparented
and across save/load cycles.

Change-Id: I65df959e4eb8d46177d59f71c8556d596e773c05
Task-number: QTCREATORBUG-17108
Reviewed-by: Marco Benelli <marco.benelli@qt.io>