Timeline: Further optimize item selection
We start from an event in the middle, rather than at the edge of the
permissible range. Then we work our way to the edges, cutting the search
short when we find that we cannot get better anymore. This reduces the
number of events that have to be checked.

Change-Id: Iba4591aa5ef71a769f26e7ee003701ffc68d3340
Task-number: QTCREATORBUG-14983
Reviewed-by: hjk <hjk@qt.io>
Reviewed-by: Joerg Bornemann <joerg.bornemann@qt.io>